# contracts/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResuAlign API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/kanban/bulk-status": {
      "post": {
        "summary": "Bulk Update Kanban Status",
        "description": "Update many kanban job statuses in one transaction.\n\n``expected_status`` acts as an optimistic lock. Replaying the same\n``idempotency_key`` returns the original result without re-applying.",
        "operationId": "bulk_update_kanban_status_api_kanban_bulk_status_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KanbanBulkStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KanbanBulkStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/master-resumes": {
      "post": {
        "summary": "Create Master Resume",
        "description": "Create a master resume with its first version.",
        "operationId": "create_master_resume_api_master_resumes_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterResumeCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Master Resumes",
        "description": "Return the current user's master resumes.",
        "operationId": "list_master_resumes_api_master_resumes_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/master-resumes/parse": {
      "post": {
        "summary": "Parse Resume Upload",
        "description": "Parse an uploaded resume file and return prefilled title/content.",
        "operationId": "parse_resume_upload_api_master_resumes_parse_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_upload_api_master_resumes_parse_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/master-resumes/{resume_id}": {
      "get": {
        "summary": "Get Master Resume",
        "description": "Return one master resume with its full version history.",
        "operationId": "get_master_resume_api_master_resumes__resume_id__get",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Master Resume",
        "description": "Append a new version to a master resume.",
        "operationId": "update_master_resume_api_master_resumes__resume_id__patch",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterResumeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Master Resume",
        "description": "Delete a master resume and its version history.",
        "operationId": "delete_master_resume_api_master_resumes__resume_id__delete",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/master-resumes/{resume_id}/diagnose": {
      "post": {
        "summary": "Diagnose Master Resume",
        "description": "Queue an independent no-JD diagnosis for one master resume.",
        "operationId": "diagnose_master_resume_api_master_resumes__resume_id__diagnose_post",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/master-resumes/{resume_id}/rollback": {
      "post": {
        "summary": "Rollback Master Resume",
        "description": "Point a master resume back at an existing version.",
        "operationId": "rollback_master_resume_api_master_resumes__resume_id__rollback_post",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resume Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterResumeRollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_parse_resume_upload_api_master_resumes_parse_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_resume_upload_api_master_resumes_parse_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KanbanBulkStatusRequest": {
        "properties": {
          "job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Job Ids"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "expected_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Status"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "job_ids",
          "status"
        ],
        "title": "KanbanBulkStatusRequest"
      },
      "KanbanBulkStatusResponse": {
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "updated": {
            "type": "integer",
            "title": "Updated",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/KanbanBulkStatusResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "title": "KanbanBulkStatusResponse"
      },
      "KanbanBulkStatusResultItem": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "updated": {
            "type": "boolean",
            "title": "Updated",
            "default": false
          },
          "status": {
            "type": "string",
            "enum": [
              "updated",
              "not_found",
              "conflict",
              "error"
            ],
            "title": "Status",
            "default": "updated"
          },
          "job": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "KanbanBulkStatusResultItem"
      },
...
      "MasterResumeCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "MasterResumeCreateRequest"
      },
      "MasterResumeRollbackRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "MasterResumeRollbackRequest"
      },
      "MasterResumeUpdateRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MasterResumeUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}